preogress: refactor
