+++ client/src/common/searchResultsContainer/index.js
@@ -1,0 +1,12 @@
+const { Box } = require("@mui/material")
+
+export const SearchResultsContainer = () => {
+    return (
+        <Box sx={{
+            width: '100%',
+            height: '100%',
+        }}>
+
+        </Box>
+    )
+}
